+++ src/main/java/com/happy/delivery/application/common/menu/MenuService.java
@@ -1,6 +1,6 @@
-package com.happy.delivery.application.menu;
+package com.happy.delivery.application.common.menu;
 
-import com.happy.delivery.application.menu.result.MenuResult;
+import com.happy.delivery.application.common.menu.result.MenuResult;
 import java.util.List;
 
 /**
