+++ src/main/java/com/happy/delivery/domain/common/menu/Menu.java
@@ -1,12 +1,9 @@
-package com.happy.delivery.application.menu.command;
+package com.happy.delivery.domain.common.menu;
 
 /**
- * MenuViewCommand.
- * layer에 맞는 dto 이름 사용, command partten을 많이 사용해서,
- * 각 계층별 api가 사용하는 dto가 다르기 때문에 dto를 분리.
+ * Menu.
  */
-public class MenuViewCommand {
-
+public class Menu {
   private Long menuId;
 
   private String menuName;
@@ -16,9 +13,9 @@ public class MenuViewCommand {
   private Long menuPrice;
 
   /**
-   * MenuViewCommand Constuctor.
+   * Menu Constructor.
    */
-  public MenuViewCommand(Long menuId, String menuName, String menuDetail, Long menuPrice) {
+  public Menu(Long menuId, String menuName, String menuDetail, Long menuPrice) {
     this.menuId = menuId;
     this.menuName = menuName;
     this.menuDetail = menuDetail;
@@ -39,5 +36,15 @@ public class MenuViewCommand {
 
   public Long getMenuPrice() {
     return menuPrice;
+  }
+
+  @Override
+  public String toString() {
+    return "Menu{" +
+        "menuId=" + menuId +
+        ", menuName='" + menuName + '\'' +
+        ", menuDetail='" + menuDetail + '\'' +
+        ", menuPrice=" + menuPrice +
+        '}';
   }
 }
