+++ src/main/java/com/happy/delivery/domain/common/menu/repository/MenuRepository.java
@@ -1,6 +1,6 @@
-package com.happy.delivery.domain.user.repository;
+package com.happy.delivery.domain.common.menu.repository;
 
-import com.happy.delivery.domain.user.Menu;
+import com.happy.delivery.domain.common.menu.Menu;
 import java.util.List;
 
 /**
