+++ src/main/java/com/happy/delivery/domain/common/restaurant/Restaurant.java
@@ -1,4 +1,4 @@
-package com.happy.delivery.domain.restaurant;
+package com.happy.delivery.domain.common.restaurant;
 
 /**
  * Restaurant.
@@ -8,16 +8,16 @@ public class Restaurant {
   private final Long id;
   private final String name;
   private final String category;
-  private final String addressCode;
+  private final String addressDetail;
 
   /**
    * Restaurant Constructor.
    */
-  public Restaurant(Long id, String name, String category, String addressCode) {
+  public Restaurant(Long id, String name, String category, String addressDetail) {
     this.id = id;
     this.name = name;
     this.category = category;
-    this.addressCode = addressCode;
+    this.addressDetail = addressDetail;
   }
 
   public Long getId() {
@@ -32,8 +32,8 @@ public class Restaurant {
     return category;
   }
 
-  public String getAddressCode() {
-    return addressCode;
+  public String getAddressDetail() {
+    return addressDetail;
   }
 
   @Override
@@ -42,7 +42,7 @@ public class Restaurant {
         "id=" + id +
         ", name='" + name + '\'' +
         ", category='" + category + '\'' +
-        ", addressCode='" + addressCode + '\'' +
+        ", addressDetail='" + addressDetail + '\'' +
         '}';
   }
 }
